refactor: remove unused styles and components from leaderboard page
