+++ leaderboard.html
@@ -37,8 +37,6 @@ porkyproductions.github.io
     <meta property="og:image" content="https://demotivator.web.app/assets/dmv-logo-a72e79b2.png" />
     <link rel="shortcut icon" href="./src/img/dmv-logo.png" type="image/x-icon">
     <title>Leaderboard | (de)Motivator</title>
-    <link rel="stylesheet" href="./src/styles/scss/darkMode.scss">
-    <link rel="stylesheet" href="./src/styles/scss/bootstrap.scss">
     <link rel="stylesheet" href="./node_modules/bootstrap-icons/font/bootstrap-icons.scss">
 
     <link rel="preconnect" href="https://fonts.googleapis.com">
